[FLINK-10816][cep] Fix LockableTypeSerializer.duplicate() to consider deep duplication of element serializer
This closes #7049.
